Calypsoify: Remove the user meta mechanism
This is a start at removing the code which relies on, and sets, the
Calypsoify user meta value. This means that the Calypsoify styling won't
be applied unless the `calypsoify` query parameter is set.

On more investigation, this probably breaks some of the flows, so it
might be best to wait until Calypsoify isn't required anymore

To get past the pre-commit hooks I needed to upgrade the PHP standards
package and the XML config. One of the rules was throwing an error.