Fix POT file creation
Currently, testing the stderr output for `wp i18n make-pot` is
incorrect. If an error occurs an exception is thrown.

If we wish to display the success or warning messages of
the `wp i18n make-pot` command, we can do so using the
info(), warning(), or success() command. I chose to minimize
these messages.